Support for disabling the 'alloptions' caching in WordPress Core
This is a frequent cause of race conditions that can easily lead to data loss and corruption. If you are running multiple web servers with time in sync(!) it's very easy to get corrupted options with long running processes and cron tasks. For a more detailed look at this problem see https://core.trac.wordpress.org/ticket/25623